pyproject.toml
[project]
name = "discord-message-counter"
version = "0.1.0"
description = "A Discord bot that tracks and counts messages sent by each member in a server"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "ollama>=0.1.0",
    "aiohttp>=3.8.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.ruff]
# Use default line length (88 characters)
line-length = 88

# Exclude common directories
exclude = [
    ".venv",
    "venv",
    ".git",
    "__pycache__",
    ".pytest_cache",
]

[tool.ruff.lint]
# Enable pycodestyle, pyflakes, and isort
select = ["E", "F", "I"]

# Ignore line-too-long for strings (E501)
ignore = ["E501"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"
